[BOOK-76] refector: datastore 모듈을 api/impl로 분리
