+++ core/datastore/api/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    alias(libs.plugins.booket.android.library)
+}
+
+android {
+    namespace = "com.ninecraft.booket.core.datastore.api"
+}
