+++ core/datastore/api/src/main/kotlin/com/ninecraft/booket/core/datastore/api/datasource/TokenPreferencesDataSource.kt
@@ -1,0 +1,11 @@
+package com.ninecraft.booket.core.datastore.api.datasource
+
+import kotlinx.coroutines.flow.Flow
+
+interface TokenPreferencesDataSource {
+    val accessToken: Flow<String>
+    val refreshToken: Flow<String>
+    suspend fun setAccessToken(accessToken: String)
+    suspend fun setRefreshToken(refreshToken: String)
+    suspend fun clearTokens()
+}
