+++ core/datastore/impl/src/main/kotlin/com/ninecraft/booket/core/datastore/impl/util/DataStoreUtil.kt
@@ -1,0 +1,16 @@
+package com.ninecraft.booket.core.datastore.impl.util
+
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.emptyPreferences
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
+import java.io.IOException
+
+fun Flow<Preferences>.handleIOException(): Flow<Preferences> =
+    catch { exception ->
+        if (exception is IOException) {
+            emit(emptyPreferences())
+        } else {
+            throw exception
+        }
+    }
