--- core/datastore/src/main/kotlin/com/ninecraft/booket/core/datastore/security/CryptoManager.kt
@@ -1,74 +1,0 @@
-package com.ninecraft.booket.core.datastore.security
-
-import android.security.keystore.KeyGenParameterSpec
-import android.security.keystore.KeyProperties
-import android.util.Base64
-import java.security.KeyStore
-import javax.crypto.Cipher
-import javax.crypto.KeyGenerator
-import javax.crypto.SecretKey
-import javax.crypto.spec.IvParameterSpec
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class CryptoManager @Inject constructor() {
-    private val cipher = Cipher.getInstance(TRANSFORMATION)
-    private val keyStore = KeyStore
-        .getInstance("AndroidKeyStore")
-        .apply {
-            load(null)
-        }
-
-    private fun getKey(): SecretKey {
-        val existingKey = keyStore
-            .getEntry(KEY_ALIAS, null) as? KeyStore.SecretKeyEntry
-        return existingKey?.secretKey ?: createKey()
-    }
-
-    private fun createKey(): SecretKey {
-        return KeyGenerator
-            .getInstance(ALGORITHM)
-            .apply {
-                init(
-                    KeyGenParameterSpec.Builder(
-                        KEY_ALIAS,
-                        KeyProperties.PURPOSE_ENCRYPT or
-                            KeyProperties.PURPOSE_DECRYPT,
-                    )
-                        .setBlockModes(BLOCK_MODE)
-                        .setEncryptionPaddings(PADDING)
-                        .setRandomizedEncryptionRequired(true)
-                        .setUserAuthenticationRequired(false)
-                        .build(),
-                )
-            }
-            .generateKey()
-    }
-
-    fun encrypt(plainText: String): String {
-        cipher.init(Cipher.ENCRYPT_MODE, getKey())
-        val iv = cipher.iv
-        val encryptedBytes = cipher.doFinal(plainText.toByteArray())
-        val combined = iv + encryptedBytes
-        return Base64.encodeToString(combined, Base64.NO_WRAP)
-    }
-
-    fun decrypt(encodedText: String): String {
-        val combined = Base64.decode(encodedText, Base64.NO_WRAP)
-        val iv = combined.copyOfRange(0, IV_SIZE)
-        val encrypted = combined.copyOfRange(IV_SIZE, combined.size)
-        cipher.init(Cipher.DECRYPT_MODE, getKey(), IvParameterSpec(iv))
-        val decryptedString = String(cipher.doFinal(encrypted))
-        return decryptedString
-    }
-
-    companion object {
-        private const val KEY_ALIAS = "secret"
-        private const val ALGORITHM = KeyProperties.KEY_ALGORITHM_AES
-        private const val BLOCK_MODE = KeyProperties.BLOCK_MODE_CBC
-        private const val PADDING = KeyProperties.ENCRYPTION_PADDING_PKCS7
-        private const val TRANSFORMATION = "$ALGORITHM/$BLOCK_MODE/$PADDING"
-        private const val IV_SIZE = 16 // AES IV는 항상 16바이트
-    }
-}
