+++ settings.gradle.kts
@@ -27,7 +27,8 @@ include(
     ":core:common",
     ":core:data:api",
     ":core:data:impl",
-    ":core:datastore",
+    ":core:datastore:api",
+    ":core:datastore:impl",
     ":core:designsystem",
     ":core:model",
     ":core:network",
